Refactored DynamicMetadata, added helper
