+++ src/Core/SourceImage.php
@@ -3,6 +3,7 @@
 namespace Rokka\Client\Core;
 
 use Rokka\Client\Core\DynamicMetadata\DynamicMetadataInterface;
+use Rokka\Client\DynamicMetadataHelper;
 
 /**
  * Represents the metadata of an image.
@@ -142,11 +143,9 @@ class SourceImage
         // Rebuild the DynamicMetadata associated to the current SourceImage
         if (isset($data['dynamic_metadata'])) {
             foreach ($data['dynamic_metadata'] as $name => $metadata) {
-                $metaClass = self::getDynamicMetadataClassName($name);
-                if (class_exists($metaClass)) {
-                    /** @var DynamicMetadataInterface $metaClass */
-                    $meta = $metaClass::createFromJsonResponse($metadata, true);
-                    $dynamic_metadata[$name] = $meta;
+                $metadata = DynamicMetadataHelper::buildDynamicMetadata($name, $metadata);
+                if ($metadata) {
+                    $dynamic_metadata[$name] = $metadata;
                 }
             }
         }
@@ -165,23 +164,5 @@ class SourceImage
             new \DateTime($data['created']),
             $data['link']
         );
-    }
-
-    /**
-     * Returns the Dynamic Metadata class name from the API name.
-     *
-     * @param string $name The Metadata name from the API
-     *
-     * @return string The DynamicMetadata class name, as fully qualified class name
-     */
-    public static function getDynamicMetadataClassName($name)
-    {
-        // Convert to a CamelCase class name.
-        // See Symfony\Component\Serializer\NameConverter\CamelCaseToSnakeCaseNameConverter::denormalize()
-        $camelCasedName = preg_replace_callback('/(^|_|\.)+(.)/', function ($match) {
-            return ('.' === $match[1] ? '_' : '').strtoupper($match[2]);
-        }, $name);
-
-        return 'Rokka\Client\Core\DynamicMetadata\\'.$camelCasedName;
     }
 }
